E-Commerce-Scraper-Suite/requirements.txt
requests
beautifulsoup4
psycopg2
SQLAlchemy
PyYAML
pytz